fix the bug in which file size was not bieng valiated in backend and alsi ad the file validation in frontend too for avatar and movie and also improve code of updtading and removing avatar

+++ backend/src/providers/upload/file-upload.ts
@@ -1,14 +1,12 @@
 import { Injectable } from '@nestjs/common';
 import { v4 as uuidv4 } from 'uuid';
-import { BadRequestException } from '@nestjs/common';
-import { Request } from 'express';
+import { InternalServerErrorException } from '@nestjs/common';
 import * as fs from 'fs';
 import * as path from 'path';
 
 @Injectable()
 export class $File {
   private readonly baseUrl = '/uploads/';
-  private readonly maxFileSize = 5000; // in bytes
 
   //   method to get url for a file that uploaded in public folder
   getPublicFileUrl(filename: string) {
@@ -18,44 +16,69 @@ export class $File {
     return `${process.env.APP_URL}${process.env.PORT}${this.baseUrl}public/${filename}`;
   }
 
+  getPublicFilePathAndName(filename: string) {
+    const uniqueName: string = uuidv4() + path.extname(filename);
+    return { path: `./uploads/public/${uniqueName}`, name: uniqueName };
+  }
+
   getPublicFilePath(filename: string) {
     return `./uploads/public/${filename}`;
   }
 
-  //   method to upload file in public folder
-  publicUpload(req: Request, file: Express.Multer.File, callback: any) {
-    const allowedFileTypes = ['image/jpeg', 'image/jpg', 'image/png'];
-
-    if (!allowedFileTypes.includes(file.mimetype)) {
-      return callback(new BadRequestException('Invalid file type'), false);
+  async checkExistenceOfFile(filePath: string): Promise<boolean> {
+    try {
+      await fs.promises.access(filePath);
+      return true;
+    } catch (error) {
+      return false;
     }
+  }
 
-    if (file.size > this.maxFileSize) {
-      return callback(
-        new BadRequestException('File size exceeds the limit'),
-        false,
+  //   method to upload file
+  async fileUpload(file: Express.Multer.File, filePath: string) {
+    try {
+      const writeStream = fs.createWriteStream(filePath);
+
+      await new Promise((resolve, reject) => {
+        writeStream.write(file.buffer, (error) => {
+          if (error) {
+            reject(error);
+          } else {
+            resolve(true);
+          }
+        });
+      });
+
+      writeStream.end();
+    } catch (error) {
+      throw new InternalServerErrorException(
+        'Got problems at backend while saving your file',
       );
     }
-
-    // If validation passes, save the file
-    const uniqueSuffix = uuidv4();
-    const extension = path.extname(file.originalname);
-    const filename = uniqueSuffix + extension;
-
-    callback(null, filename);
   }
 
   // remove file
-  deleteFile(filePath: string) {
-    return new Promise((resolve, reject) => {
-      fs.unlink(filePath, (error) => {
-        if (error) {
-          console.log(error);
-          reject(false);
-        } else {
-          resolve(true);
-        }
+  async deleteFile(filePath: string) {
+    try {
+      const isExist = await this.checkExistenceOfFile(filePath);
+      if (!isExist) {
+        return true;
+      }
+
+      await new Promise((resolve, reject) => {
+        fs.unlink(filePath, (error) => {
+          if (error) {
+            reject(error);
+          } else {
+            resolve(true);
+          }
+        });
       });
-    });
+      return true;
+    } catch (error) {
+      throw new InternalServerErrorException(
+        'Got problems at backend while deleting your file',
+      );
+    }
   }
 }
